Decouple HCA Compressor and utils
